DOMAIN: Moved basket class
Moved basket class to orders folder and removed unnecessary usings
