+++ BelissimoCloneWPF.Domain/Commons/Auditable.cs
@@ -1,4 +1,10 @@
-﻿namespace BelissimoCloneWPF.Domain.Commons
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BelissimoCloneWPF.Domain.Commons
 {
     public abstract class Auditable
     {
