+++ BelissimoCloneWPF.Domain/Entities/Baskets/Basket.cs
@@ -1,8 +1,8 @@
 ﻿using BelissimoCloneWPF.Domain.Commons;
 
-namespace BelissimoCloneWPF.Domain.Entities.Orders
+namespace BelissimoCloneWPF.Domain.Entities.Baskets
 {
-    public sealed class Basket : Auditable
+    public sealed class Basket: Auditable
     {
         public int UserId { get; set; }
         public int TotalPrice { get; set; }
