+++ BelissimoCloneWPF.Domain/Entities/Users/User.cs
@@ -8,5 +8,6 @@ namespace BelissimoCloneWPF.Domain.Entities.Users
         public string  Name { get; set; }
         public string PhoneNumber { get; set; }
         public UserRole Role { get; set; }
+
     }
 }
