+++ BelissimoCloneWPF.Domain/Enums/UserRole.cs
@@ -1,4 +1,10 @@
-﻿namespace BelissimoCloneWPF.Domain.Enums
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BelissimoCloneWPF.Domain.Enums
 {
     public class UserRole
     {
